test: split test_private into two test files
